make lib.rs pure routing
